+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0
+
+### Added
+
+- Types for some of Bento HTML elements (`<bento-embed />` and `<bento-sidebar />`)
+- Types for new document event `bento-onComponentVisibilityChange`
+
 ## 1.0.0
 
 ### Added